+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Web Component interface.
 - Images for ui.
 - Router class.
+- Deploy script.
 
 ## [0.1.0] - 2021-01-25